chore(frontend/context): 🎨 move theme provider to context
