+++ app/frontend/src/components/mode-toggle.tsx
@@ -1,6 +1,3 @@
-import { Moon, Sun } from "lucide-react";
-
-import { useTheme } from "@/components/theme-provider";
 import { Button } from "@/components/ui/button";
 import {
 	DropdownMenu,
@@ -8,6 +5,8 @@ import {
 	DropdownMenuItem,
 	DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
+import { useTheme } from "@/context/theme-provider";
+import { Moon, Sun } from "lucide-react";
 
 export function ModeToggle() {
 	const { setTheme } = useTheme();
